Move mockVapProfile into fixtures folder
