+++ src/platform/forms-system/test/js/components/ContactInfo.unit.spec.jsx
@@ -3,7 +3,7 @@ import { Provider } from 'react-redux';
 import { expect } from 'chai';
 import { render, waitFor, fireEvent } from '@testing-library/react';
 
-import vapProfile from '../../../../user/profile/vap-svc/tests/e2e/mockVapProfile.json';
+import vapProfile from '../../../../user/profile/vap-svc/tests/fixtures/mockVapProfile.json';
 
 import { $, $$ } from '../../../src/js/utilities/ui';
 import ContactInfo from '../../../src/js/components/ContactInfo';
