+++ src/platform/user/profile/vap-svc/tests/fixtures/mockVapProfile.json
